Enable public fields without getters to be displayed
Enable public fields without getters to be displayed in the list.
Previously, when there was not getter method, the field would not have a value in the list, even if it was public. Forms would still work because they use the Symfony2 Form mechanisms that look for getters first and then for public fields.