Update standard_layout for non-logged in users
The sonata_top_nav_menu (add_block + user_block) was drawn all the time for my login page after the update, so I just added a quick check to remove it.

My login page extends base_template as my project only consists of an admin section, and therefor my login page is now slightly weird as it contains a blank sidebar. Should I be making a new base_template or improving standard_layout so that it looks normal for users that are not logged in? (Remove the 250px margin and sidebar if user is not logged in?)