Extend money_widget and percent_widget
Input fields are streched to 100% width in the new SonataAdminBundle, but this
doesn't take into account any extra text such as percentage signs or currency
symbols, causing either the symbol or the input field to be pushed to the next
line.

Using Bootstrap's input-group we can prevent this, and also render the symbol
in a nice little box.

The money widget respects the money_pattern variable, so the currency symbol
will be rendered before or after the input, depending on the locale or the
given money_pattern.