Update standard_layout.html.twig
When resizing to xsmall devices in bootstrap the breadcrumb is breaking the mobile responsive layout (all of the actions, etc fall over to the next line and no longer line up with the background). 
Ideally the breadcrumb div would be updated to sit within a variable width container that would allow the user to pan left and right with the mouse/their finger to read the entire breadcrumb on mobile devices. If I get a chance I can look at integrating this, but for the time being I suggest wrapping it up in <div class="hidden-xs"></div> so that it doesn't appear broken to the user on xsmall devices.