Add blocks inside the table and th tags in base_list.html.twig
The additional blogs allow adding attributes and classes to these tags without needing to copy a much larger section of the template.

Some javascript libraries (FooTable.js for example) use data attributes to set properties of the columns or table. This can easily be done by overriding the new blocks.