Update CRUDController.php
CRUDController uses $this->getRequest() and $this->get('request') to work with the current request object.

These changes are trying to change (fix?) that. The request object is injected into the action methods by type hinting it in the parameters list (this is a BC break). Also the controller is configured (the configure() method is called) in the newly introduced setRequest() method, not when the container is injected (this should not cause any problems and is not a BC break).