Update form_types.rst
If no admin class for the linked model class is defined no options are rendered in the form. It is really hard to debug this. I only found the solution when I found the following issue: See issue: https://github.com/sonata-project/SonataAdminBundle/issues/2012

Adding this note to the documentation could save people hours of debugging.