Load jquery ui after the boostrap js to fix tooltip conflict
Admin LTE enables the bootstrap tooltip by default. But if the jquery ui is loaded after the bootstrap js, the bootstrap tooltip plugin is replaced by jquery's tooltip plugin. Because of different parameters it's wrongly enabled on all items with `title` attribute.
This commit fixes the tooltip conflict, the bootstrap tooltip plugin is prefered.