Fixed Scope Crossing Injection
Fixed error

Scope Crossing Injection detected: The definition "sonata.admin.builder.orm_datagrid" references the service "sonata.admin.builder.filter.factory" which belongs to another scope hierarchy. This service might not be available consistently. Generally, it is safer to either move the definition "sonata.admin.builder.orm_datagrid" to scope "prototype", or declare "container" as a child scope of "prototype". If you can be sure that the other scope is always active, you can set the reference to strict=false to get rid of this error.