Update the_form_view.rst
Lost about an hour trying to figure out why new symfony3 is yelling on 'property' option on entity when tried to complete tutorial. Figured out that it is deprecated but it took some time that now we should use 'choice_label'. Hope this note will help to other who try to learn Sonata on new symfony3.
'sonata_type_model' has this option on FormMapper, but not on DatagridMapper as I understood.