Remove default template loading on exception
Why would anyone want to rely on that? Maybe there is a valid reason,
but it does not appear in the comments or in the commit history. The
best way to find out is to remove this "feature". If anyone reverts this
commit, please make sure to add a comment to explain this.