Use form type names instead of FQCN (reverts #3709)
My previously merged change turned out to not fix, but break the admin
panel. While there can be no garantee that this commit "fixes" anything
instead of breaking it, the change of getting a working admin panel will
be significantly increased after this PR. This because it reverts things
back to the state they were for the past 5 years (without any bug report
about it being broken).