Add a second argument to batchActionDelete()
In order to perform some additional actions before you delete some element in the list view, the preDelete method is provided by CRUDController.php. By overriding this method you can for example check some conditions and prevent deletion by returning a Response object. However if you delete multiple elements in list view the batchActionDelete will be called and there is no "preBatchActionDelete" method available.
 
Of course there is the "preBatchAction" method, but it has some limitations: 
1) it's declared in admin class and not in CRUDController
2) it can't prevent deletion, bacuse the returning value will not be taken into account
3) It is general and will be called regardless of actual batch action (delete or some other custom action)

We can override the batchActionDelete method which is located in CRUDController, but this method takes only one parameter: $query, so we have no array with all IDs (like in preBatchAction)
It is strange, because if we create some custom batch actions it will receive two arguments: $query and $request. We can than get from the request all necessary information (e.g. list of chosen IDs) and depending on it change the $query or prevent deletion by returning new Response object.

Actually both of them, batchActionDelete and custom batchAction<MyAction> will be called by the same function(line 457), so both of them actually receive the second parameter: $request.

My suggestion is just to declare the second argument in batchActionDelete, because it will already be passed to it.

Thanks