Bump SonataCoreBundle version constraint
People that upgrade the admin bundle in order to get Symfony 3.3
compatibility might not be aware that the core bundle is compatible with
it only since 3.4.0, and not think of updating it too. This change is
not pure in the sense that this incompatibility should be expressed in
the core bundle itself, but it should spare us some support issues.