Remove unused field
Hiding is not enough. If you are using AJAX, and your field is required,
you may end up with an empty select tag, with a required attribute that
is never filled with options, because there are currently 2 tags, one
for each version of Select2, and only one of them is filled.
This commit removes the other one instead of just hiding it,
that way it can't interfere with the other.
Fixes #4486, fixes #4261