Add the `set translation_domain` inside the else clause
The line with the `set translation_domain` must be called in the IF clause and in the ELSE clause. The IF will happen when the item being translated is inside a treeview, which means that the label must be translated against the translation_domain option. The ELSE will happen when on_top is true, which means that the label shoule be translated against the label_catalogue option.