Do not override final Controller methods
Since Symfony 3.4 most methods in the ControllerTrait are considered to be final, including the render() method which we override.

While we can rename "our" method, this will introduce a BC issue with child classes that still use render().

This is a workaround for that.

Because Symfony < 3.3 did not have the ContainerTrait (which is required for this solution) we polyfill it.