Inherited templates from parent services handling.
If you defined a service and overriden templates
(i.e. with `setTemplate` or `setTemplates`) they
were not propagated to child services.

This fixes the bug but is not perfect because
`setTemplate` and `setTemplates` calls are
normally removed from the service definition.

Here we don't remove them from inherited parent
because otherwise, other children would have an
incomplete template set.