improvments on fetching current class (#4824)
There are different ways to get the current class of an entity. It can
either be the class of the active subclass (request query parameter
'subclass'), or the class of the subject, which the modelmanager would
serve for the current id. At least if both does not match the class is
the class passed as constructor parameter.
This is the behavior the getClass() method is created for and this is
the behavior the doc block in the interface describes. So the
change takes care to do so and also fixes fetching and computing the
subclass from its hash map. Other improvements:
- let it fail, when subclass does not exist in mapping array.
- getObject should use injected class
- break the loop by not using getter for subject in getClass()
- remove some more direct property usages