Trigger autoload to ensure BC
A class that extends one of ours, or implements one of our interface
but still uses legacy types in the signature of the methods it overrides
or implements will cause a crash at compile time: type hints in
signature do not cause autoloading, and thus the class alias will not be
created. To fix this, we trigger the autoload of the new type, so that
no deprecation is thrown when it in turns autoloads the old type.

That issue was reported here:
https://github.com/sonata-project/SonataAdminBundle/pull/5418#issuecomment-454128691