Fix Twig template Sonata Form Type Autocomplete
Adding a new item such as an image using the autocomplete form type could result in an exception, as an array was expected via the form, but a string was returned. Also, a variable decoding a JSON string was crashing the JavaScript on the page, while it was not being used.