Add the basic doctrine metadata
To make things simpler, I suggest that we add the basic Doctrine metadata on the class level just like it has been done on the class properties.

I also added a reference to the Doctrine documentation page.

Adding the $id might be nice, but as least now if you execute the `doctrine:schema:create` command, you will get a nice error message to tell you that something is wrong in your entity. Until now, the only message on a fresh install was that no entity was found.. how frustrating.