Prevent installing Doctrine bundle 2
We rely on a "doctrine" service defined in that bundle, which means we
have a dependency on the bundle that cannot be seen by just examining
php code like composer-require-checker does.
That dependency is optional though, because not everyone is using the
ORM or ACLs. Using the conflict section is the best way to prevent
installation of version 2 while still not requiring installation of
version 3.
The error might be quite easy to fix, but the conflict I added should
stay, and just be modified to ">=3" when the error is fixed.

See:
- https://github.com/doctrine/DoctrineBundle/blob/d1f8a28fb45b6ff0b59f88df393575a4f89c3b66/Resources/config/dbal.xml#L67
- https://github.com/sonata-project/SonataAdminBundle/blob/3ab3cca373f76304149aa2ba0ef7da7d34c998b6/src/Resources/config/commands.xml#L20-L24

Refs #5757