Apply latest cs fixes
The removal of &MockObject is a known bug in PHP-CS-Fixer, but I am not
sure we can afford to wait for it to be fixed and released. Since this
was the only occurence of this in the project, I think it is acceptable
to drop it for now.

See https://github.com/FriendsOfPHP/PHP-CS-Fixer/issues/4913
See https://github.com/FriendsOfPHP/PHP-CS-Fixer/pull/4915