Fixed automatically marking actions, batch and actions field as virtual
The check for the field type is moved so the right type is used when the user passes `null` as the type and the type is guessed by the storage implementation.

The type of the _action field is now guessed in the ListMapper because it is not a model property and otherwise each model manager must perform this type guess.