Avoid toggle filter button to be shown when the filter is hidden
When an admin is child, a filter pointig to its parent is automatically added at
`AbstractAdmin::buildDatagrid()` using the value from `getParentAssociationMapping()`
as name. This filter is rendered as hidden, but since it's considered active, the
toggle button used to remove the filter is shown.
This change uses the "show_filter" option to determine if this button must be shown.