Remove passing admin_pool to twig templates in next major version
Since sonata-project/admin-bundle 3.83, Pool is just used as an
Admin registry and for global options has been replaced by
SonataConfiguration which holds those options. To access to
these options from Twig, there is a global Twig sonata_config
variable.

So there is no need to pass an instance of Pool to Twig
templates anymore, the pool is not supposed to be used
directly in templates.