Add RouteCollectionInterface
This interface was added in 4.x branch, this commit imports it
removing type and return declarations to make it compatible with
RouteCollection which implements this interface.

In sonata 4 AbstractAdmin::configureRoutes() method receives a
RouteCollectionInterface, importing this interface in 3.x would
allow users to update their code and define this method as in 4
easing the upgrading process.