Add preBatchAction to Admin Extension and add a pre_batch_action event
This adds a preBatchAction() method to the AbstractAdminExtension and
AdminExtensionInterface (BC). This enables the AdminEventExtension to
dispatch a BatchActionEvent of type pre_batch_action. Both the extension
and the event provide points to hook into within your own code, just
like this is possible for pre/post persist/update/delete.

Fixes #7516 "Implement an event for batch actions"
Relates to #6550 "Delete Event isn't triggered in Batch delete"