Fix error on getNewInstance() when parent object property in uninitialised. (#7705)
When creating a new object instance, the admin will check if there is a
parent object that needs to be appended. The appendParentObject() method
will check the current value of the property that holds the parent
object. But if this property is typed and not yet initialised, Symfony's
PropertyAccess will throw an AccessException. This fix catches that
exception and continues as if the current value is `null`.